Cria módulos para funções
